Create main.py
Yes, you can write a Python program to retrieve the location information for a given IP address using these geolocation services. To do this, you would need to use a library or API provided by the service to make a request and retrieve the location data. The exact code would depend on the specific library or API you are using.